requirements.txt
# Install these in a clean Python 3.10 virtual environment for this project only.
# CreditSense AI does not require TensorFlow or FAISS. If pip reports conflicts
# involving packages like `tensorflow-intel` or `faiss-cpu`, the active
# interpreter already contains unrelated packages and should be recreated.
Django==4.2.7
djangorestframework==3.15.2
pandas==2.2.3
numpy==1.24.3
scikit-learn==1.7.2
shap==0.48.0
lime==0.2.0.1
psycopg2-binary==2.9.10
